// src/lib/api/generated/models/updateServiceAccount.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Hubuum REST API
 * OpenAPI documentation for the Hubuum REST service.
 * OpenAPI spec version: 0.0.1
 */

/**
 * Mutable fields on a service account.
 */
export interface UpdateServiceAccount {
  /** @nullable */
  description?: string | null;
  /** @nullable */
  owner_group_id?: number | null;
}